django/src/requirements.txt
Django==5.2
djangorestframework==3.15.2
drf-spectacular==0.27.2
django-cors-headers==4.6.0
psycopg2-binary==2.9.10
django-environ==0.11.2
daphne==4.1.2
celery[redis]==5.4.0
redis==5.2.1
google-api-python-client==2.163.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
factory-boy==3.3.1
